adds dummy index for auth debugging
